move tests into raft_test.go per reviewer request
